+++ src/main/scala/it/unibo/pps/evoldino/prolog/PrologInScala_olds.scala
@@ -1,27 +1,42 @@
-package it.unibo.pps.evoldino.model
+package it.unibo.pps.evoldino.prolog
 
 import alice.tuprolog
+import alice.tuprolog.*
 import it.unibo.pps.evoldino.controller.engine.EngineConstants
-import it.unibo.pps.evoldino.model.dinosaur.Dinosaur
+import it.unibo.pps.evoldino.model.{createDemoDinosaur, dinosaur}
+import it.unibo.pps.evoldino.model.dinosaur.*
 import it.unibo.pps.evoldino.model.disaster.Disaster
 
 import scala.collection.mutable.ListBuffer
-import alice.tuprolog.*
-
 import scala.language.postfixOps
 
-object prologFunextends7 extends App:
+object prologFunextends10 extends App:
   import it.unibo.pps.evoldino.prolog.PrologInScala.{*, given}
 
-  val engine: Term => LazyList[SolveInfo] = mkPrologEngine("""
+  val myDinoList = createDemoDinosaur()
 
-    dino_position(70, 30).
-    dino_position(45, 45).
-    dino_position(20, 20).
-    dino_position(21, 31).
-    dino_position(71, 31).
-    dino_position(29, 29).
-    dino_position(30, 30).
+  val dinX = new ListBuffer[scala.Int]
+  val dinY = new ListBuffer[scala.Int]
+
+  //stampare con una funzione scala style i dinosauri :)
+  myDinoList.zipWithIndex.map {
+    case (dino, count) =>
+      println(s"$count is $dino")
+      dinX += dino.coordinates._1
+      dinY += dino.coordinates._2
+  }
+
+  //i Disasteri per semplicità li ho inventati direttamente dentro Prolog
+
+  val engine: Term => LazyList[SolveInfo] = mkPrologEngine(s"""
+
+    dino_position(${dinX(0)},${dinY{0}}).
+    dino_position(${dinX(1)},${dinY{1}}).
+    dino_position(${dinX(2)},${dinY{2}}).
+    dino_position(${dinX(3)},${dinY{3}}).
+    dino_position(${dinX(4)},${dinY{4}}).
+    dino_position(${dinX(5)},${dinY{5}}).
+
 
     dis_position(60, 20).
     dis_position(45, 45).
@@ -40,6 +55,11 @@ object prologFunextends7 extends App:
   //val res = engine("same(X, Y)") foreach (println(_))
   //engine("same(X, Y)") foreach (println(_))
 
+  //NOTARE CHE UNA IS A LIST OF solveinfo e l altra di term
+
+  //IO HO USATO TUPROLOG IN JAVA, MA PENSO SIA MEGLIO NUOVA VERSIONE IN KOTLIN !!
+  //PENSO CHE FORSE MEGLIO SE LA STRINGA DELL' ENGINE SI SPOSTI FUORI NEL FILE .pl e si chiami poi
+
   val DinoFromProlog1 = new ListBuffer[SolveInfo]
   val DinoFromProlog2 =  new ListBuffer[Term]
 
@@ -49,6 +69,7 @@ object prologFunextends7 extends App:
   val DinolistP1 = DinoFromProlog1.toList
   val DinolistP2 = DinoFromProlog2.toList
 
+  //questo e' per prendere e stampare i dinosauri da prolog
   print("print list dinosaur" + "\n")
   for (i <- DinolistP1)
     print("list: " + i + "\n")
@@ -57,16 +78,16 @@ object prologFunextends7 extends App:
   for (i <- DinolistP2)
     print("list sol: " + i + "\n")
 
+  //questo è per prendere e stampare i disastri da prolog
   val DisFromProlog =  new ListBuffer[Term]
-
-  val DisProlog = engine("dis_position(X,Y)") foreach ( DisFromProlog += _.getSolution )
-
+  val DisProlog: Unit = engine("dis_position(X,Y)") foreach ( DisFromProlog += _.getSolution )
   val DisList = DisFromProlog.toList
 
   print("print list disaster sol" + "\n")
   for (i <- DisList)
     print("list sol: " + i + "\n")
 
+  //questo passa la funzione all engine e poi viene stampata
   val ListFromPrologSameP = new ListBuffer[Term]
   val sameP = engine("same(X,Y)") foreach ( ListFromPrologSameP += _.getSolution )
 
@@ -78,15 +99,16 @@ object prologFunextends7 extends App:
 
   val lp2 = new ListBuffer[SolveInfo]
 
+  //stessa cosa di sopra ma qui uso la funzione di TuProlog per prendere solo una variabile
   val SameP2 = engine("same(X,Y)") foreach ( lp2 += _)
   val lp2List= lp2.toList
 
-  for((el,i) <- lp2List.zipWithIndex)
-    print("element: " + i + " is " + "\n"
-    + "X:" + el.getVarValue("X") + "\n"
-    + "Y:" + el.getVarValue("Y") +"\n")
-
-
+  lp2List.zipWithIndex.foreach {
+    case (el, i) =>
+      print("element: " + i + " is " + "\n"
+        + "X:" + el.getVarValue("X") + "\n"
+        + "Y:" + el.getVarValue("Y") + "\n")
+  }
